libs/utils/analysis: add FREQUENCY module
This patch move the CPUs/Clusters frequency residency analysis code, originally
part of the trace_analysis.py module, to be a new TraceAnalysis module whith
dynamic registration via the AnalysisRegister.
